Split comparisons.cpp into files for each specific comparison callable. This works around the old "object file has too many sections" bug on MinGW. In making this fix it also became clear that some of the kernel headers were missing some includes. This fixes that as well.
