+++ include/dynd/kernels/total_order_kernel.hpp
@@ -6,6 +6,7 @@
 #pragma once
 
 #include <dynd/kernels/base_kernel.hpp>
+#include <dynd/kernels/base_strided_kernel.hpp>
 #include <dynd/types/fixed_string_type.hpp>
 
 namespace dynd {
